requirements.txt
alembic==1.7.5
APScheduler==3.6.3
celery[redis]
Flask==2.0.2
Flask-Cors==3.0.10
Flask-JWT==0.3.2
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
flask-restx==0.5.1
Flask-SocketIO==5.1.1
Flask-SQLAlchemy==2.5.1
Flask-SSE==1.0.0
gevent==21.8.0
gunicorn==20.1.0
psycopg2-binary==2.9.2
PyJWT==1.4.2
python-binance==1.0.15
python-dotenv==0.19.2
requests==2.26.0
websockets==9.1
Werkzeug==2.0.2
wheel==0.37.0